connecting the dots and updating the schema
