+++ validator/db/constants.py
@@ -33,19 +33,10 @@ ASSIGNED_MINERS = "assigned_miners"
 TASK_ID = "task_id"
 ACCOUNT_ID = "account_id"
 MODEL_ID = "model_id"
-MODEL_FILENAME = "model_filename"
-DS_URL = "ds_url"
-DS_ID = "ds_id"
-FIELD_SYSTEM = "field_system"
-FIELD_INSTRUCTION = "field_instruction"
-FIELD_INPUT = "field_input"
-FIELD_OUTPUT = "field_output"
-FORMAT = "format"
-NO_INPUT_FORMAT = "no_input_format"
+DS = "ds"
 STATUS = "status"
 HOURS_TO_COMPLETE = "hours_to_complete"
 TEST_DATA = "test_data"
-SYNTHETIC_DATA = "synthetic_data"
 TRAINING_DATA = "training_data"
 MINER_SCORES = "miner_scores"
 CREATED_AT = "created_at"
@@ -58,6 +49,18 @@ IS_ORGANIC = "is_organic"
 TIMES_DELAYED = "times_delayed"
 ASSIGNED_MINERS = "assigned_miners"
 TASK_TYPE="task_type"
+
+# Text Tasks Table Columns
+FIELD_SYSTEM = "field_system"
+FIELD_INSTRUCTION = "field_instruction"
+FIELD_INPUT = "field_input"
+FIELD_OUTPUT = "field_output"
+FORMAT = "format"
+NO_INPUT_FORMAT = "no_input_format"
+SYNTHETIC_DATA = "synthetic_data"
+
+# Image Tasks Table Columns
+MODEL_FILENAME = "model_filename"
 
 # Submissions Table Columns
 SUBMISSION_ID = "submission_id"
